restructure detail allocation folder
