+++ src/views/reports/allocation/detail/infoLabel.js
